refactor: breaking out API into their respective files
